requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
plyer==2.1.0
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
watchdog==3.0.0
werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
